split test_logout.py: shared scenarios vs Python-specific tests
Revert the validator When/Then exemption (9f08ec68) — the orphan checker
is working as intended per Adam's feedback: it forces you to think about
whether a test belongs in shared or is language-specific.

Instead, split test_logout.py into two files:
- test_logout.py: 4 methods matching shared/session/logout.feature scenarios
  (token cleanup, query rejection, idempotent close, concurrent close)
- test_close.py: 13 Python-specific tests (wrapper config,
  auto-detection, keep-alive remapping, retry, atexit auto-cleanup)

The new file doesn't match any shared feature by name, so the orphan
checker skips it entirely. The assert_logout_request_format helper
moved to the new file (only used there).
